+++ packages/sdk-router/src/sdk.ts
@@ -328,7 +328,7 @@ class SynapseSDK {
     poolAddress: string,
     amount: BigNumber
   ): Promise<{
-    amounts: Record<string, { value: BigNumber; index: number }>
+    amounts: Array<{ value: BigNumber; index: number }>
     routerAddress: string
   }> {
     const router: SynapseRouter = this.synapseRouters[chainId]
@@ -336,11 +336,13 @@ class SynapseSDK {
       poolAddress,
       amount
     )
-    const poolTokens = await router.routerContract.poolTokens(poolAddress)
-    const amountsOut: Record<string, { value: BigNumber; index: number }> = {}
-    poolTokens.map((token, index) => {
-      amountsOut[token.token] = { value: amounts[index], index }
-    })
+    const amountsOut: Array<{ value: BigNumber; index: number }> = amounts.map(
+      (respAmount, index) => ({
+        value: respAmount,
+        index,
+      })
+    )
+
     return {
       amounts: amountsOut,
       routerAddress: router.routerContract.address,
@@ -351,20 +353,12 @@ class SynapseSDK {
     chainId: number,
     poolAddress: string,
     amount: BigNumber,
-    token: string
+    poolIndex: number
   ): Promise<{
     amount: { value: BigNumber; index: number }
     routerAddress: string
   }> {
     const router: SynapseRouter = this.synapseRouters[chainId]
-
-    let poolIndex = 0
-    const poolTokens = await router.routerContract.poolTokens(poolAddress)
-    poolTokens.map((poolToken, index) => {
-      if (poolToken.token === token) {
-        poolIndex = index
-      }
-    })
 
     const outAmount = await router.routerContract.calculateWithdrawOneToken(
       poolAddress,