+++ packages/synapse-interface/components/TokenInput.tsx
@@ -1,12 +1,15 @@
 import { useCallback, MouseEvent } from 'react'
 import InteractiveInputRow from './InteractiveInputRow'
 import { displaySymbol } from '@utils/displaySymbol'
-import { Token } from '@types'
+import { PoolToken, PoolUserData, Token } from '@types'
 import { cleanNumberInput } from '@utils/cleanNumberInput'
+import { formatUnits } from '@ethersproject/units'
+import { BigNumber } from '@ethersproject/bignumber'
 
-const TokenInput = ({
+export const DepositTokenInput = ({
   token,
   balanceStr,
+  rawBalance,
   inputValueStr,
   onChange,
   chainId,
@@ -14,6 +17,7 @@ const TokenInput = ({
 }: {
   token: Token
   balanceStr: string
+  rawBalance: BigNumber
   inputValueStr: string
   onChange: (v: string) => void
   chainId: number
@@ -24,7 +28,10 @@ const TokenInput = ({
   const onClickMax = useCallback(
     (e: MouseEvent<HTMLButtonElement>) => {
       e.preventDefault()
-      onChange(balanceStr)
+
+      const adjustedValue = formatUnits(rawBalance, token.decimals[chainId])
+
+      onChange(adjustedValue)
     },
     [onChange, balanceStr, token]
   )
@@ -52,4 +59,59 @@ const TokenInput = ({
     </div>
   )
 }
-export default TokenInput
+
+export const WithdrawTokenInput = ({
+  poolUserData,
+  token,
+  balanceStr,
+  inputValueStr,
+  onChange,
+  chainId,
+  address,
+}: {
+  poolUserData?: PoolUserData
+  token: Token
+  balanceStr: string
+  inputValueStr: string
+  onChange: (v: string) => void
+  chainId: number
+  address: string
+}) => {
+  const symbol = displaySymbol(chainId, token)
+
+  const onClickMax = useCallback(
+    (e: MouseEvent<HTMLButtonElement>) => {
+      e.preventDefault()
+      const adjustedValue = formatUnits(
+        poolUserData?.lpTokenBalance,
+        token.decimals[chainId]
+      )
+
+      onChange(adjustedValue)
+    },
+    [onChange, balanceStr, token]
+  )
+
+  return (
+    <div className="items-center">
+      <div className="w-full">
+        <InteractiveInputRow
+          title={symbol}
+          isConnected={address !== undefined}
+          balanceStr={balanceStr}
+          onClickBalance={onClickMax}
+          value={inputValueStr}
+          placeholder={'0.0000'}
+          onChange={(e) => onChange(cleanNumberInput(e.target.value))}
+          // disabled={inputValueStr == ''}
+          disabled={false}
+          showButton={false}
+          icon={token?.icon?.src}
+          token={token}
+          isPending={false}
+          onClickEnter={() => {}}
+        />
+      </div>
+    </div>
+  )
+}