+++ packages/synapse-interface/utils/actions/getPoolData.ts
@@ -1,5 +1,4 @@
 import { Zero, One } from '@ethersproject/constants'
-import { calculateExchangeRate } from '@utils/calculateExchangeRate'
 import { getEthPrice, getAvaxPrice } from '@utils/actions/getPrices'
 import {
   commifyBnToString,
@@ -16,6 +15,8 @@ import {
 import { fetchBalance, fetchToken } from '@wagmi/core'
 import { PoolTokenObject, Token, PoolUserData, PoolData } from '@types'
 import { BigNumber } from 'ethers'
+
+import { getVirtualPrice } from './getPoolFee'
 
 const getBalanceData = async ({
   pool,
@@ -34,7 +35,7 @@ const getBalanceData = async ({
   const lpTotalSupply =
     (
       await fetchToken({
-        address: `0x${lpTokenAddress.slice(2)}`,
+        address: lpTokenAddress as `0x${string}`,
         chainId,
       })
     )?.totalSupply?.value ?? Zero
@@ -45,12 +46,14 @@ const getBalanceData = async ({
 
     const rawBalance = (
       await fetchBalance({
-        address: `0x${address.slice(2)}`,
+        address: address as `0x${string}`,
         chainId,
-        token: `0x${token.addresses[chainId].slice(2)}`,
+        token: token.addresses[chainId] as `0x${string}`,
       })
     )?.value
 
+    // TODO: this is to support virtual price calcs, which needs to get updated
+    // as a contract call
     const balance = rawBalance.mul(
       BigNumber.from(10).pow(18 - token.decimals[chainId])
     )
@@ -73,6 +76,7 @@ const getBalanceData = async ({
       poolTokenSum = poolTokenSum.add(balance)
     }
   }
+
   return {
     tokenBalances,
     poolTokenSum,
@@ -80,6 +84,7 @@ const getBalanceData = async ({
     lpTotalSupply,
   }
 }
+
 export const getPoolData = async (
   chainId: number,
   pool: Token,
@@ -99,17 +104,16 @@ export const getPoolData = async (
 
   const lpTokenAddress = pool?.addresses[chainId]
 
-  const { tokenBalances, poolTokenSum, lpTokenBalance, lpTotalSupply } =
-    await getBalanceData({
+  const { tokenBalances, lpTokenBalance, lpTotalSupply } = await getBalanceData(
+    {
       pool,
       chainId,
       address: user ? address : poolAddress,
       lpTokenAddress,
-    })
+    }
+  )
 
-  const virtualPrice = lpTotalSupply.isZero()
-    ? MAX_BN_POW
-    : calculateExchangeRate(lpTotalSupply, 18, poolTokenSum, 18)
+  const virtualPrice = await getVirtualPrice(poolAddress, chainId)
 
   const ethPrice = prices?.ethPrice ?? (await getEthPrice())
   const avaxPrice = prices?.avaxPrice ?? (await getAvaxPrice())
@@ -146,6 +150,7 @@ export const getPoolData = async (
       tokens: poolTokensMatured,
       lpTokenBalance,
       lpTokenBalanceStr: formatBNToString(lpTokenBalance, 18, 4),
+      nativeTokens: pool.nativeTokens,
     }
   }